feat: Update code samples (2/2).
